improve monster movement
Until now a spawned monster's random movement probability was biased towards
the spawn point. This resulted in the monsters' random walks converging at the
spawn. Now movement is qually distributed in the monster's boundaries. If a
monster would walk out of its radius, the walking direction is reversed for
that move. With the L_inf metric, we are using, this ensures that the monster
stays inside its range.