Only players should talk to non-players
If an NPC talks to an NPC, the global thisNPC changes during an NPC entrypoint. This may cause problems for the outer entrypoint if thisNPC is to be used again in the same call. For a proper fix all NPC entrypoints need to store the former value of thisNPC and reset it after they are done. Here it has to be taken into account that entrypoints may crash and that thisNPC has to be reset regardless. Such a fix will be implemented in the future.